moving analog to its own file
