+++ common/hw.h
@@ -154,12 +154,16 @@ inline static void evsys_config(u8 channel, u8 source, u8 user) {
 
 #define EVSYS_EVD(N) ((N)<=7 ? (1<<((N) + 8)) : (1 << (24 + (N) - 8)))
 
+// analog.c
+void adc_init(u8 channel);
+void dac_init(u8 channel);
+uint16_t analog_read(Pin p);
+void analog_write(Pin p, u16 val);
+
 // clock.c
 void gclk_enable(uint32_t id, uint32_t src, uint32_t div);
 void clock_init_usb(u8 clk_system);
 void clock_init_crystal(u8 clk_system, u8 clk_32k);
-void adc_init(u8 channel);
-void dac_init(u8 channel);
 
 // dma.c
 #define DMA_DESC_ALIGN __attribute__((aligned(16)))
